Migrate "leboncoin feedback" job (#297)
