+++ api/src/models/mission.ts
@@ -217,6 +217,8 @@ schema.index({ organizationName: 1 });
 schema.index({ organizationRNA: 1 });
 schema.index({ organizationClientId: 1 });
 schema.index({ organizationReseaux: 1 });
+schema.index({ leboncoinStatus: 1 });
+schema.index({ jobteaserStatus: 1 });
 
 schema.index({ geoPoint: "2dsphere" });
 schema.index({ "addresses.geoPoint": "2dsphere" });
